+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::getMultiVendorProductById`
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::createMultiVendorProduct`
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::updateMultiVendorProduct`
+ - Added `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationIdFromUserId`
 
 ## 1.32.0