+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## Next release
 
+- Added `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationFromUserId`
+
 ### 1.33.0
 
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::getMultiVendorProductById`
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::createMultiVendorProduct`
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::updateMultiVendorProduct`
- - Added `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationIdFromUserId`
 
 ## 1.32.0