+++ CHANGELOG.md
@@ -7,6 +7,8 @@
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::getMultiVendorProductById`
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::createMultiVendorProduct`
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::updateMultiVendorProduct`
+ - Added `\Wizaplace\SDK\Company\CompanyService::updateFile`
+ - Added `\Wizaplace\SDK\Company\CompanyService::deleteFile`
 
 ## 1.32.0